+++ examples/build-lnmp/outputs.tf
@@ -1,0 +1,3 @@
+output "nginx_url" {
+  value = "${element(alicloud_eip.default.*.ip_address, 1)}:80/test.php"
+}