+++ terraform/examples/alicloud-oss-bucket-object/outputs.tf
@@ -1,8 +1,0 @@
-output "bucket-new" {
-  value = "${alicloud_oss_bucket.bucket-new.id}"
-}
-
-output "content" {
-  value = "${alicloud_oss_bucket_object.content.id}"
-}
-