+++ CHANGELOG.md
@@ -4,6 +4,7 @@ IMPROVEMENTS:
 
 - Output application attribution service block (([#435](https://github.com/alibaba/terraform-provider/pull/435)))
 - Add kubernetes example (([#436](https://github.com/alibaba/terraform-provider/pull/436)))
+- Add connections output for kubernetes cluster (([#437](https://github.com/alibaba/terraform-provider/pull/437)))
 
 ## 1.9.1 (March 30, 2018)