+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 ## 1.9.2 (Unreleased)
 
+IMPROVEMENTS:
+
+- Update example (([#457](https://github.com/alibaba/terraform-provider/pull/457)))
+
 BUG FIXES:
 
+- Fix deleting forward table not found error (([#457](https://github.com/alibaba/terraform-provider/pull/457)))
 - Fix deleting slb listener error (([#439](https://github.com/alibaba/terraform-provider/pull/439)))
 - Fix creating vswitch error (([#439](https://github.com/alibaba/terraform-provider/pull/439)))