+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 IMPROVEMENTS:
 
+- *New Resource*: _alicloud_ots_table_(([#442](https://github.com/alibaba/terraform-provider/pull/442)))
 - Update example (([#457](https://github.com/alibaba/terraform-provider/pull/457)))
+- Remove terraform/example (([#458](https://github.com/alibaba/terraform-provider/pull/458)))
 
 BUG FIXES: